trading_system/High_frequency_crypto_tradin/requirements.txt
# High-Frequency Crypto Trading System Requirements
# =================================================

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning
tensorflow>=2.14.0

# Alpaca Trading
alpaca-py>=0.13.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance
numba>=0.58.0

# Data quality
pyarrow>=14.0.0
